+++ Zadanie 1/Task/src/Task.cpp
@@ -4,7 +4,7 @@ Task::Task() : m_pj(0), m_rj(0), m_qj(0), m_taskId(0)
 {
 }
 
-Task::Task(double pj, double rj, double qj, int id) : m_pj(pj), m_rj(rj), m_qj(qj), m_taskId(id)
+Task::Task(int pj, int rj, int qj, int id) : m_pj(pj), m_rj(rj), m_qj(qj), m_taskId(id)
 {
 }
